Fix a bunch of cppcheck stuff Also upgrade & swap to multi-file nlohmann

+++ core/source/sav/Sav7.cpp
@@ -152,7 +152,7 @@ u8 Sav7::badges(void) const
     u8 ret        = 0;
     for (size_t i = 0; i < sizeof(badgeBits) * 8; i++)
     {
-        ret += badgeBits & (1 << i) ? 1 : 0;
+        ret += (badgeBits & (u32(1) << i)) ? 1 : 0;
     }
     return ret;
 }
